[vocabulary] Reworking VDEXVocabularyMixin to not open a the soucre vdex file and load on each term lookup; now the vdex file is loaded and parsed on init only.
Renaming BaseVDEXVocabulary to TreeVDEXVocabulary.

Unconvoluting VocabularyTextField lookup of an ITreeVocabulary utility (that is always the same as the instance of what is set as self.vocabulary anyway!). Using direct vocabulary reference in descriptor as opposed to going via a registered utility on ITreeVocabulary, as will be done when descriptor classes are created dynamically); 

Making FlatVDEXVocabulary code more explicit and informative; elaborating comments on the class.