[nav tree link ids] Adding an id atrribute on each <a> element of each entry in the navigation tree (used in admin section), to facilitate predicatble selection of these links by UI testing frameworks such as selenium.
Each id value is guaranteed to be unique for the tree, and a differentiator between trees may be specified if more than one such navigation is ever to be included on a same page.

Replacing a variety of inelegant, and sometimes incorrect, tests for a "string" value with isinstance tests on builtin type basestring.

Other misc src cleanup.