[declarative] Making workspace add parlaimentary content menu items and workspace add views to be dynamically registered for types with an enabled workspace feature; adding ui.feature module.
Deleting previously hard-wired declarations in ui/menu.zcml and ui/views.zcml.

Note: if a type enables the "workspace" feature then it also gains the workspace UI components for adding that type; if this is not desireable, then qualifiers on the "workspace" to make such distinctions are necessary.

This change makes it possible to add a new type to bungeni entirely from configuration (note: derived / extended / sub-object container attributes not yet supported from config), with the possibility to create instances of the type in the workspace.